Created a service to wrap the configuration details of http requests.
