+++ source/services/authentication/authentication.service.ts
@@ -1,6 +1,6 @@
 import {Injectable} from 'angular2/core';
-import {Http, Response, Headers, RequestOptions} from 'angular2/http';
 import {Observable} from 'rxjs/observable';
+import { RequestService } from '../request/request.service';
 
 export interface IUser {
 	username: string;
@@ -16,26 +16,24 @@ export class AuthenticationService {
 	isAuthenticated: boolean = false;
 	loggedInUser: IUser;
 	
-	constructor(private http: Http) {}
+	constructor(private http: RequestService) {}
 	
 	login(credentials: ICredentials): Observable<IUser> {
-		let body = JSON.stringify(credentials);
-		let headers = new Headers({ 'Content-Type': 'application/json' });
-		let options = new RequestOptions({ headers: headers });
-		return this.http.post('/users/login', body, options)
-					.map(res =>  {
-						this.loggedInUser = res.json().data;
-						console.log(credentials.username + ' is now logged in');
+		return this.http.post('/users/login', credentials)
+					.map(user =>  {
+						this.loggedInUser = user;
+						console.log(user.username + ' is now logged in');
 						this.isAuthenticated = true;
 						return this.loggedInUser;
-					})
-					.catch(this.handleError);
+					});
 	}
 	
-	private handleError (error: any) {
-		// In a real world app, we might send the error to remote logging infrastructure
-		let errMsg = error._body || 'Server error';
-		console.error(errMsg); // log to console instead
-		return Observable.throw(errMsg);
+	register(credentials: ICredentials): Observable<IUser> {
+		return this.http.post('/users/register', credentials)
+					.map(user =>  {
+						this.loggedInUser = user;
+						this.isAuthenticated = true;
+						return this.loggedInUser;
+					});
 	}
 }
